maint: use ASAN-poisoning more carefully
The ASAN-poisoning instituted by commit v2.21-14-g1555185 was
incomplete, since the poisoned tail of the read buffer could well
be the target of a legitimate follow-on read.  To accommodate that,
we must unpoison each such region just before beginning fillbuf's
read loop.
* src/grep.c [HAVE_ASAN] (asan_poison): Define.
(clear_asan_poison): Define.
(fillbuf): Clear before reading, since we are likely to read
into memory that was poisoned on the preceding iteration.
* tests/two-files: New file, to test for this.
* tests/Makefile.am (TESTS): Add it.