grep: output "." file name in diagnostic
This is bug C as reported by David Grayson in:
http://bugs.gnu.org/16444#18
This bug occurs only in obscure circumstances, and I didn't see
how to write a reasonable test case for it.
* src/grep.c (filename_prefix_len): Remove, replacing with ...
(omit_dot_slash): New static var.  All uses of the former replaced
with uses of the latter.
(grepdirent): Don't add 2 if the filename is just ".".