tests/backref-multibyte-slow:: avoid false positive
* tests/backref-multibyte-slow: When redirecting the "fast" LC_ALL=C
run's output to /dev/null, we got an artificially low timing (of 0),
due to grep's own stdout-vs-/dev/null optimization.  With an initial
timing of 0 on that first run, the derived timeout for the UTF-8 run
(which redirects to a file) would be a mere 1 second.  The fix: also
redirect that first run's output to a file, not to /dev/null.