grep: improve dfasearch storage management
This patch is mostly refactoring, with a bit of performance tweaking.
It is done in preparation for a fix for Bug#24009.
* src/dfasearch.c (patterns): Now of type struct re_pattern_buffer *
instead of an anonymous struct pointer, since there is no longer
any need to keep regs here.  All uses changed.
(GEAcompile): Use patlim instead of a hard-to-follow "total".
Use x2nrealloc to avoid potential O(N**2) reallocation algorithm.
Initialize just the pattern members that need clearing.
(EGexecute): Put regs into a static variable, as this code did
before 2001-02-18, as there is no need to have a separate set of
regs for each pattern.  Explain the "Q@#%!#" comment better.