grep: speed up multiple word matching
grep uses its KWset matcher for multiple word matching, but that is
very slow when most of the parts matched to a pattern are not words.
So, if the first match to a pattern is not a word, use the grep matcher
to match for its line.

Note that when START_PTR is set, the grep matcher uses the regex matcher
which is very slow to match words.  Therefore, we use the grep matcher
when only START_PTR is NULL.

* src/kwsearch.c (Fexecute): If an initial match is incomplete because
not on a word boundary, use the grep matcher to find a matching line.