grep: avoid false -Fw match in non-UTF8 multibyte locales
For example, this command would erroneously print its input line:
  echo ab | LC_CTYPE=ja_JP.eucjp grep -Fw b
This arose when the "memrchr" search for a preceding newline failed:
in that case, MB_START was not adjusted and was initially the same
as BEG, so wordchar_prev mistakenly returned 0.
* src/kwsearch.c (Fexecute): Set MB_START also when there is no
preceding newline.
* NEWS (Bug fixes): Mention it.
* tests/mb-non-UTF8-word-boundary: New file. Test for the bug.
* tests/Makefile.am (TESTS): Add it.
Reported by NIDE, Naoyuki in https://bugs.gnu.org/38223.